pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "fmri-pipeline"
version = "0.1.0"
description = "A modular, YAML-configured pipeline for multi-scale functional brain network analysis"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Sara Pradhan"}
]
keywords = ["fMRI", "neuroimaging", "functional connectivity", "reproducibility"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
dependencies = [
    "numpy>=1.21",
    "scipy>=1.7",
    "pandas>=1.3",
    "nibabel>=3.2",
    "nilearn>=0.9",
    "scikit-learn>=1.0",
    "matplotlib>=3.4",
    "seaborn>=0.11",
    "pyyaml>=5.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=3.0",
]

[tool.setuptools.packages.find]
where = ["src"]
